no need for .vue extension in the import statements. also moved icon base component to icon folder
